Zozo

Description

Library for constructing, exploring, and visualizing regular expressions and their NFA form. Focusing heavily on the Brzozowski derivative but also Thompson's construction.

Documentation

data Regex

A simple regex lang

Constructors

Nil
The empty set

Eps
The empty string, \( \epsilon \)

Sym Char
A symbol as a singleton set

Union Regex Regex
Union of language sets \[ \Lambda \vee \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Inter Regex Regex
Intercept of language sets \[ \Lambda \wedge \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Comp Regex
Complement of language sets \[ \neg \Gamma \hspace{10pt} \Gamma \subseteq \Sigma^* \]

Conc Regex Regex
All possible concatenations \[ \{a\}\{b\} := \{a, b, ab, ba\} \]

Star Regex
Kleene star operator

Regex constructors

(<|>) :: Regex -> Regex -> Regex

Synonym for the Union constructor

(<^>) :: Regex -> Regex -> Regex

Synonym for the Inter constructor

(***) :: Regex -> Regex -> Regex

Synonym for the Conc constructor

Instances

Derivatives

derive :: Char -> Regex -> Regex

Defining the (Brzozowski) derivative of a language R with respect to the string \(u \in \Sigma^*\) to be \[ u^{-1} L = \left\{ v : uv \in L \right\} \]

as :: Regex

many :: Char -> Regex

as' :: Regex