Zozo

Description

Library for constructing, exploring, and visualizing regular expressions and their NFA form. Focusing heavily on the Brzozowski derivative but also Thompson's construction (eventually).

Language definition

data Regex

A simple regex lang

Constructors

Nil
The empty set

Eps
The empty string, \( \epsilon \)

Sym Char
A symbol as a singleton set

Union Regex Regex
Union of language sets \[ \Lambda \vee \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Inter Regex Regex
Intercept of language sets \[ \Lambda \wedge \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Comp Regex
Complement of language sets \[ \neg \Gamma \hspace{10pt} \Gamma \subseteq \Sigma^* \]

Conc Regex Regex
All possible concatenations

Star Regex
Kleene star operator

Regex constructors

(<|>) :: Regex -> Regex -> Regex infixl 6

Synonym for the Union constructor

(<^>) :: Regex -> Regex -> Regex infixl 6

Synonym for the Inter constructor

(***) :: Regex -> Regex -> Regex infixl 6

Synonym for the Conc constructor

neg :: Regex -> Regex

Synonym for the Comp constructor

(\\) :: Regex -> Regex -> Regex infixl 6

Language set difference \[ \Lambda \backslash \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Specialized constructors

r :: String -> Regex

Treat string as concatenation of symbol regexes

r"foo" = Sym f <> Sym o <> Sym o

some :: Regex -> Regex

Synonym for Star constructor That is, zero or more

many :: Regex -> Regex

One or more

alphabet :: String -> Regex

Fold string over language union

numericChar = alphabet ['0'..'9']

unit :: Bool -> Regex

\[ \mathbb{1}_x = \begin{cases} \epsilon \quad & x = \top \\ \varnothing \quad & \text{otherwise} \\ \end{cases} \]

nu :: Regex -> Regex

Auxiliary function to derive where \[ \nu(R) = \begin{cases} \epsilon \quad & \epsilon \in R\\ \varnothing \quad & \text{otherwise} \\ \end{cases} \] Equivalent to unit . (Eps mem)

Derivatives

derive :: String -> Regex -> Regex

Defining the (Brzozowski) derivative of a language \(L\) with respect to the string \(u \in \Sigma^*\) to be \[ u^{-1} L = \left\{ v : uv \in L \right\} \]

(^-) :: String -> Regex -> Regex infixl 5

Infix derive with regex symbol input

Evaluation

rreduce :: String -> Regex -> Regex

Continue reducing a regex

eval :: Regex -> Set String

Evaluate regex to a set of strings

match :: String -> Regex -> Bool

Check if regex matches. Essentially checking if the derivative with respect to a string is equal to the empty string \(\epsilon\)

(?) :: String -> Regex -> Bool infixl 5

Infix synonym for match

Regex sets

alpha :: Regex

Alphabetic characters

numeric :: Regex

Numeric characters

alphaNum :: Regex

Alpha-numeric characters

space :: Regex

Single space character

spaces :: Regex

One or more spaces

whiteSpace :: Regex

Any whitespace character

anyChar :: Regex

Any ASCII character