Zozo

Description

Library for constructing, exploring, and visualizing regular expressions and their NFA form. Focusing heavily on the Brzozowski derivative but also Thompson's construction.

Language definition

data Regex

A simple regex lang

Constructors

Nil
The empty set

Eps
The empty string, \( \epsilon \)

Sym Char
A symbol as a singleton set

Union Regex Regex
Union of language sets \[ \Lambda \vee \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Inter Regex Regex
Intercept of language sets \[ \Lambda \wedge \Gamma \hspace{10pt} \Lambda,\Gamma \subseteq \Sigma^* \]

Comp Regex
Complement of language sets \[ \neg \Gamma \hspace{10pt} \Gamma \subseteq \Sigma^* \]

Conc Regex Regex
All possible concatenations

Star Regex
Kleene star operator

Regex constructors

(<|>) :: Regex -> Regex -> Regex infixl 5

Synonym for the Union constructor

(<^>) :: Regex -> Regex -> Regex infixl 5

Synonym for the Inter constructor

(***) :: Regex -> Regex -> Regex infixl 5

Synonym for the Conc constructor

neg :: Regex -> Regex

Synonym for the Comp constructor

Specialized constructors

r :: String -> Regex

Treat string as concatenation of symbol regexes

some :: Regex -> Regex

Synonym for Star constructor That is, zero or more

many :: Regex -> Regex

One or more

unit :: Bool -> Regex

\[ \mathbb{1}_x = \begin{cases} \epsilon \quad & x = \top \\ \varnothing \quad & \text{otherwise} \\ \end{cases} \]

nu :: Regex -> Regex

Auxiliary function to derive where \[ \nu(R) = \begin{cases} \epsilon \quad & \epsilon \in R\\ \varnothing \quad & \text{otherwise} \\ \end{cases} \] Equivalent to unit . (Eps mem)

Derivatives

mem :: Regex -> Regex -> Bool

Set membership

derive :: String -> Regex -> Regex

Defining the (Brzozowski) derivative of a language R with respect to the string \(u \in \Sigma^*\) to be \[ u^{-1} L = \left\{ v : uv \in L \right\} \]

(^-) :: String -> Regex -> Regex infixl 4

Infix derive with regex symbol input

Evaluation

rreduce :: String -> Regex -> Regex

Continue reducing a regex

eval :: Regex -> Set String

Evaluate regex to a set of strings